fix: code improvements

+++ src/libsyncengine/syncpal/syncpalworker.h
@@ -83,7 +83,7 @@ class SyncPalWorker : public ISyncWorker {
          */
         bool isLocalItemInSyncWithDb(const SyncPath &localAbsolutePath, std::optional<NodeId> &outLocalNodeId);
 
-        ExitInfo ensureBlackListIsPropagated(int trial = 0);
+        ExitInfo ensureBlackListIsPropagated(int16_t trial = 0);
 
         void ensureMinimumPermission();
